Add `dantil.skipFileInLocationRetrieval()`
Adds `dantil.skipFileInLocationRetrieval()`, which marks the module in
which the method is invoked for `dantil.getModuleCallerLocation()` to
skip when searching the call stack. Extends
`dantil.getModuleCallerLocation()` to support this method.

This is useful for using `dantil.getModuleCallerLocation()` to include
a method's invocation location in an error message, though that
location is deep within the call stack relative to the error's
generation location. I.e., the error is caught several modules deep
from the invocation to which the error applies, as opposed to being one
module deep.